Database almost OK

+++ app/src/main/java/hu/balygaby/projects/cyclepower/connectivity/InputData.java
@@ -5,7 +5,14 @@ import android.location.Location;
 public interface InputData {
     int INPUT_VALID = 1;
 
-    void transmitBicycleData(int validity, double wheelRpm, double pedalRpm, double wheelRotation);
+    /**
+     * Callback when a new Bluetooth message arrives.
+     * @param validity Validity of the Bluetooth data.
+     * @param wheelRpm Current wheel rpm.
+     * @param pedalRpm Current pedal rpm.
+     * @param wheelRotation Current count of wheel rotations. This value can only increase with time.
+     */
+    void transmitBicycleData(int validity, double wheelRpm, double pedalRpm, long wheelRotation);
 
     /**
      * Callback when we have a new location.
